- client: fix function prototypes for CUDA detection.
    client/
        coproc_detect.cpp

svn path=/branches/boinc_core_release_7_0a/; revision=25618

+++ client/coproc_detect.cpp
@@ -822,7 +822,7 @@ typedef int (__stdcall *CUDA_GDM)(size_t*, int);
 typedef int (__stdcall *CUDA_GDCC)(int*, int*, int);
 typedef int (__stdcall *CUDA_CC)(void**, unsigned int, unsigned int);
 typedef int (__stdcall *CUDA_CD)(void*);
-typedef int (__stdcall *CUDA_MA)(unsigned int*, unsigned int);
+typedef int (__stdcall *CUDA_MA)(unsigned int*, size_t);
 typedef int (__stdcall *CUDA_MF)(unsigned int);
 typedef int (__stdcall *CUDA_MGI)(size_t*, size_t*);
